Added VoicePlayer.js from react-voice-components
Added VoicePlayer.js from react-voice-components as create-react-app wouldn't load react-voice-components (because of old or eccentric syntax)
and I did not want to eject create-react-app to access webpack.config.js.   The solution was to copy it into my src library.
Still debugging it, there are still some issues with it, but it does play text to voice.