Partially fixed issue with conditional rendering of voice player
Voice player now plays correctly when the voice state is set to null in
the select handler for the dropdown.  However, setting the voice state to null
in the handleClick for the radio button doesn't help and if the user clicks a differnt
value on the radio button without selecting from the dropdown first, the voice still won't play.